migrations TODO done

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -19,7 +19,12 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            //$table->float('commission_rate');
+
+            //Floating-point numbers can have rounding errors, and when dealing with financial calculations, it's crucial to maintain precision to avoid discrepancies.
+            // therefore it is recommended to use the decimal data type. The decimal data type provides fixed-point arithmetic, which helps avoid rounding errors associated with floating-point numbers.
+            $table->decimal('commission_rate',10,3);
+
             $table->string('discount_code');
             $table->timestamps();
         });
